Refaktorering header-logo
