--- src/ikoner/meny/NavLogoRod.less
@@ -1,14 +1,0 @@
-@import '../../../node_modules/nav-frontend-core/less/_variabler';
-@import '../../styling-variabler.less';
-
-.hovedmeny .navbar-brand {
-    padding-right: 0.25rem;
-    min-width: 97.2px;
-
-    &:focus,
-    &:active {
-        background-color: initial;
-        outline: @fokusFarge 3px solid;
-        border-radius: 4px;
-    }
-}
