+++ src/komponenter/header/header-regular/mobil/MobilMenylinje.less
@@ -28,32 +28,6 @@
         flex-wrap: wrap;
     }
 
-    &__logo {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        width: 66px;
-        height: 66px;
-
-        svg {
-            width: 62px;
-            height: 62px;
-        }
-
-        svg:hover {
-            width: 66px;
-            height: 66px;
-        }
-
-        &:hover {
-            outline: 3px solid black;
-        }
-
-        &:focus {
-            outline: 3px solid @fokusFarge;
-        }
-    }
-
     &__sok {
         margin-left: 3px;
         text-decoration: none;
